Safari-611.1.10/LayoutTests/mathml/presentation/menclose-notation-no-overlap-expected.txt

Tests that menclose with notation value as circle is sqrt(2) times the inner width/height.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS clientRect.width >= 141 is true
PASS clientRect.height >= 70 is true
PASS successfullyParsed is true

TEST COMPLETE
